user, goal, goallog delete를 store에서 전체 delete하도록 수정
