--- src/main/java/com/cyhee/rabit/service/comment/BasicCommentService.java
@@ -1,54 +1,0 @@
-
-package com.cyhee.rabit.service.comment;
-
-import java.util.Optional;
-
-import com.cyhee.rabit.model.cmm.ContentStatus;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.stereotype.Service;
-
-import com.cyhee.rabit.model.cmm.ContentType;
-import com.cyhee.rabit.exception.cmm.NoSuchContentException;
-import com.cyhee.rabit.dao.comment.CommentRepository;
-import com.cyhee.rabit.model.comment.Comment;
-
-@Service("basicCommentService")
-public class BasicCommentService implements CommentService {
-	@Autowired
-	private CommentRepository repository;
-	
-	public Page<Comment> getComments(ContentType type, Long parentid, Pageable pageable) {
-		return repository.findByTypeAndParentIdAndStatusNot(type, parentid, ContentStatus.DELETED, pageable);
-	}
-
-	public Page<Comment> getComments(Pageable pageable) {
-		return repository.findByStatusNot(ContentStatus.DELETED, pageable);
-	}
-
-	public void addComment(Comment comment) {
-		repository.save(comment);
-	}
-
-	public Comment getComment(long id) {
-		Optional<Comment> comment = repository.findById(id);
-		if(!comment.isPresent())
-			throw new NoSuchContentException(ContentType.COMMENT, id);
-		return comment.get();
-	}
-
-	public void updateComment(long id, Comment source) {
-		Comment comment = getComment(id);
-		setCommentProps(comment, source);
-	}
-
-	public void deleteComment(long id) {
-		Comment comment = getComment(id);
-		comment.setStatus(ContentStatus.DELETED);
-	}
-	
-	private void setCommentProps(Comment target, Comment source) {
-		target.setContent(source.getContent());
-	}
-}
