+++ src/main/java/com/cyhee/rabit/service/goal/GoalStoreService.java
@@ -1,6 +1,8 @@
 package com.cyhee.rabit.service.goal;
 
 import com.cyhee.rabit.model.cmm.ContentStatus;
+
+import com.cyhee.rabit.service.comment.CommentService;
 import com.cyhee.rabit.service.goallog.GoalLogService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -9,7 +11,6 @@ import org.springframework.stereotype.Service;
 
 import com.cyhee.rabit.model.cmm.ContentType;
 import com.cyhee.rabit.model.comment.Comment;
-import com.cyhee.rabit.service.comment.CommentService;
 import com.cyhee.rabit.model.goal.Goal;
 import com.cyhee.rabit.dao.goallog.GoalLogRepository;
 import com.cyhee.rabit.model.goallog.GoalLog;
@@ -18,7 +19,9 @@ import java.util.List;
 
 @Service
 public class GoalStoreService {
-	
+	@Autowired
+	GoalService goalService;
+
 	@Autowired
 	GoalLogRepository goalLogRepository;
 	
@@ -26,7 +29,12 @@ public class GoalStoreService {
 	CommentService commentService;
 
 	@Autowired
-    GoalLogService goalLogService;
+	GoalLogService goalLogService;
+
+	public void deleteGoal(long id) {
+		Goal goal = goalService.deleteGoal(id);
+		deleteAllGoalStore(goal);
+	}
 
 	public Page<GoalLog> getGoalLogs(Goal goal, Pageable pageable) {
 		return goalLogRepository.findByGoalAndStatusNot(goal, ContentStatus.DELETED, pageable);
@@ -40,7 +48,12 @@ public class GoalStoreService {
 		return commentService.getComments(ContentType.GOAL, goal.getId(), pageable);
 	}
 
+	public List<Comment> getComments(Goal goal) {
+		return commentService.getComments(ContentType.COMMENT, goal.getId());
+	}
+
 	public void deleteAllGoalStore(Goal goal) {
 		getGoalLogs(goal).forEach(gl -> goalLogService.deleteGoalLog(gl.getId()));
+		getComments(goal).forEach(cmt -> commentService.deleteComment(cmt.getId()));
 	}
 }
