+++ src/main/java/com/cyhee/rabit/web/goallog/GoalLogStoreController.java
@@ -1,5 +1,6 @@
 package com.cyhee.rabit.web.goallog;
 
+import com.cyhee.rabit.service.goallog.GoalLogService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -13,7 +14,6 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.cyhee.rabit.model.comment.Comment;
 import com.cyhee.rabit.model.goallog.GoalLog;
-import com.cyhee.rabit.service.goallog.GoalLogService;
 import com.cyhee.rabit.service.goallog.GoalLogStoreService;
 
 @RestController
