+++ src/test/java/com/cyhee/rabit/goallog/GoalLogApiTest.java
@@ -6,6 +6,7 @@ import static org.mockito.Mockito.when;
 
 import java.util.Arrays;
 
+import com.cyhee.rabit.service.goallog.GoalLogService;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,7 +20,6 @@ import org.springframework.test.web.servlet.MockMvc;
 
 import com.cyhee.rabit.cmm.ApiTestUtil;
 import com.cyhee.rabit.model.goallog.GoalLog;
-import com.cyhee.rabit.service.goallog.GoalLogService;
 import com.cyhee.rabit.web.goallog.GoalLogController;
 
 @RunWith(SpringRunner.class)
@@ -28,7 +28,7 @@ import com.cyhee.rabit.web.goallog.GoalLogController;
 public class GoalLogApiTest {
 	@Autowired
 	private MockMvc mvc;
-	@MockBean(name="basicGoalLogService")
+	@MockBean(name="goalLogService")
     private GoalLogService goalLogService;
 	
 	private static final String url = "/rest/v1/goallogs";
